add rust
